+++ src/front/js/layout.js
@@ -14,12 +14,13 @@ import { SignupRelacion } from "./pages/signup-relacion";
 import { SignupAborto } from "./pages/signup-aborto";
 import { SignupTratamientos } from "./pages/signup-tratamientos";
 import { SignupCentros } from "./pages/signup-centros";
+import { ListOfWomen } from "./pages/list-of-women";
+import { Profile } from "./pages/profile";
 
 import injectContext from "./store/appContext";
 
 import { Navbar } from "./component/navbar";
 import { Footer } from "./component/footer";
-import { ListOfWomen } from "./pages/List-of-women";
 
 //create your first component
 const Layout = () => {
@@ -71,6 +72,9 @@ const Layout = () => {
 						</Route>
 						<Route exact path="/list-of-women">
 							<ListOfWomen />
+						</Route>
+						<Route exact path="/profile">
+							<Profile />
 						</Route>
 						<Route>
 							<h1>Not found!</h1>
