+++ src/front/js/pages/signup-aborto.js
@@ -4,6 +4,7 @@ import rigoImageUrl from "../../img/rigo-baby.jpg";
 import "../../styles/home.scss";
 import { Link } from "react-router-dom";
 import ProgressBar from "react-bootstrap/ProgressBar";
+import { SignupRelacion } from "./signup-relacion";
 
 export const SignupAborto = () => {
 	return (
@@ -12,8 +13,13 @@ export const SignupAborto = () => {
 			<label>numero</label>
 			<input type="number" />
 
-			<Link to={"/signup-relacion"}>Siguiente</Link>
-			<button>Saltar pregunta</button>
+			<Link to={"/signup-relacion"}>
+				<button>Siguiente</button>
+			</Link>
+			<Link to={"/list-of-women"}>
+				<button>Saltar Cuestonario</button>{" "}
+			</Link>
+
 			{/* <progress id="file" max="100" value="40">
 				{" "}
 				40%{" "}
